Add "ble_app_template" nordic example
It was added the copy of the "ble_app_template" from the nordic example.
The folder and the project was renamed to "main_code_nrf52840".

Only the necessary files from the original nordic example were left in
the new project folder.

The project is meant to be used with the nordic the PCA10056 Board, the
140 Softdevice and the SES Development Environment.

The .gitignore file was added in order to Git doesn't detect changes
when a local user compile the code in SES.