First android BLE connection
This is the first time the "main_code_android" and the
"main_code_nrf52840" can be tested with a successful BLE connection.

The android app is using the "Template" service. So, in order to test,
open the phone app and select the template service, then press
"connect", select the "Nord" device and then the notified values should
be received by the phone.