Add functional but messy NOTIFY, READ, WRITE
The logic of the progam is as follows:
(a) The user open the template activity which is now called "DNA Amplifier"
an then press the "Connect" button.

(b) Once connected, the phone receives notifications from the "STAT"
characteristic every 10 seconds.

(c) If the user press the "read" button, then the characteristic is read
from the nRF52, this can be notice by looking at the android logcat
debugger.

(d) If the user press the "write" button, then the phone writes in the
STAT characteristic a prefixed data buffer, then the nRF52 notifies to
the phone back the value. Notice that the data buffer is {-128,0,-50,0,-1,0,
1,0,50,0,127,0} which is displayed in the phone app as {128,206,255,1,50,127},
which is correct considering unsigned/signed decimal representation.