Clean to read, write, notify STAT service
- Clean logic to display data on UI: The TextViews now are an array
of TextsViews. This was made just to have a neat code.
- Remove unnecesary UUIDs names
- Now there are just two considered services: the "Generic Access" and
the "STAT" service. The important service is the "STAT" service, however
the "Generic Access" was left in the code to understand the logic in case
we need to add another service/characteristic.
- Basically, the code checks if the "STAT" characteristic and the "Device Name"
characteristic is present in the Ble device. If they are, then it is displayed
in the phone as a connectable device. Once connected, the "STAT" characteristic
is continuosly been notified, and can be read and written by the android app.