Add status variable and minimal communication logic
- There is a new module which is in charge of storing and changing the
status of the nordic device.
- There is a minimal communication logic between the android app and the
nordic device. First, the nordic device is turned on and waits for a
command from the phone. Second, the phone connects and then the user can
press the button 'WRITE'. Third, the nordic detects that the phone wrote
on the STAT characteristic and the nordic toggles the status variable of
the command received from the phone. Fourth, if the command variable is
true, then the nordic enables the detection system task, otherwise the
task is disabled.