Add functional STAT communication logic for nRF52
This commit performs a very complete communication logic of the STAT
characteristic with the commandFromPhone and isMeasuring varibles.

The communication logic is implemented just in the nRF52 and not in the
android app. There must be a future commit where this code can work with
the main code of the nRF52 of this version. So far, this code can be
tested with the nRFconnect app.

The communication flow will be explained:
- First, the nRF52 is turned on and waits until a phone is connected via
BLE with the nRFconnect app.
- Second, once the nRFconnect app is connected, then it should enable
notifications of the STAT characteristic. Note that the STAT
characteristic can be read or written. However, only if the notification
is enabled and the user writes a byte different from zero 0x00, then the
nRF52 will start the detection system task.
- Third, if the nRF52 is measuring and then the user write 0x00 in the
nRFconnect app, then the detection system tasks will stop.
- Fourth, note that there are more conditions to avoid undesired states
in the logic of the communication and the stored states of the
commandFromPhone and isMeasuring variables. To understand this cases,
please feel free to test with the nRFconnect app and look the log
messages printed out by the nRF52 via uart.