Add functional but messy STAT logic for android
This version is just a messy but functional solution for the
STAT characteristic communication logic. Basically, it is the same
functionality that has the last commit with the nRFconnect app, but now
it is possible to use the main_android_code app.

Note that every time the user press the button "WRITE", then an internal
variable changes in the android app to change the state of the STAT
characteristic. This is not the best suited way, instead the phone app
must read the STAT characteristic from the nRF52before any change to it.

The communication logic is the same as the past commit. For further
insights, please refer to the past commit.