Update UI after Read or Notify STAT characteristic
Now, the android app updates the user interface asychronously after read
the STAT characteristic. This functionality is similar when there is a
notification (already implemented). Every time the STAT characteristic is
updated by either Read or Notify, it is emited an event that lets the user
interface to be updated asynchronously.