Add simple save file feature
Now the app can save a file with the information of the SENS service.

Once the app is connected and the user press the "START" button, then
the app listen that the nRF52 got the commandFromPhone and isMeasuring
by being notified on the STAT service. Thus, the app create a new file
named with the date of the phone in the "../DCIM/DNAamplifier" folder.

Then, every time the app receives a notification from the SENS service,
it writes a new line in the file in .csv format.

Finally, the file is closed after the user press the "STOP" button.

Note that if the app is in the background or the phone has its screen
turned off, the notification are still being received and the file is
constantly being written. Also, if the app disconnects from bluetooth
for a while but then reconnects, then all the data stored in the flash
of the nRF52 is notified in the SENS characteristic sequentially, and
as fast as possible, to the android app which successfully saves the
file.

However, if the user closes the app while a measuring process is being
performed by the nRF52, then a new file is created which is not desirable.
This issue must be faced in a future commit.