Modify save file based on isDataOnFlash from STAT service
Now, the communication logic to create a file and display the state
of the "START/STOP" button in the android app depends on the state
of the variables commandFromPhone and isDataOnFlash which are stored
in the STAT characteristic of the nRF52.

Note that the isMeasuring variable is not considered for the logic,
however it serves to the user to know if the measuring process is
running. For that reason, in the nRF52 code and in the android code
the variables related with isMeasuring are not longer used and the
isDataOnFlash is being used insted.

It's relevant to point out that in the nRF52 the general state change
of the STAT characteristic is as follows:
- Initial State:
commandFromPhone = 0. isMeasuring = 0. isDataOnFlash = 0.
- Once user on the android app press the "START" button:
commandFromPhone = 1. isMeasuring = 1. isDataOnFlash = 1.
- Once user press "STOP" button:
commandFromPhone = 0. isMeasuring = 1. isDataOnFlash = 1.
commandFromPhone = 0. isMeasuring = 0. isDataOnFlash = 1.
commandFromPhone = 0. isMeasuring = 0. isDataOnFlash = 0.
Refer to the serial communication of the nRF52 for more detail.

Also,  note that for a future version the isMeasuring can be stopped
internally by the nRF52 once the measuring process reaches a timeout.
In this case the state of the STAT characteristic in the nRF52 would be:
commandFromPhone = 1. isMeasuring = 0. isDataOnFlash = 1.
And therefore the logic in the android app should change a bit
for displaying the "START/STOP" button (maybe another changes are
neccesary too).

The flow is as follows:
First, the nRF52 and android app must be connected via BT. Then the
user press the "START" button, and a file is created and being written
every time a notification of the "SENS" characteristic is received.
In this state can happen four things:
(a) User press the "STOP" button
(b) User turn-offs the screen or the app is left in the background
(c) User press the "DISCONNECT" button
(d) User closes the app
If (a), then the created file is closed.
If (b), then the created file is hold open being written every time
a new "SENS" characteristic is received. Eventually, the user will go
back to the app and will press the "STOP" button, so the (a) situation
will happen.
If (c), then the created file is hold open but doensn't receive BT
notifications, so is not being written. However, eventually the user
will press the "CONNECT" button and the app will receive notifications
sequentially as fast as possible from the "SENS" characteristic. After
that, the user should press the "STOP" button, so the (a) situation
will happen.
If (d), then the created file is closed. However, eventually the user
will press the "CONNECT" button and the app will receive notifications
sequentially as fast as possible from the "SENS" characteristic. Here,
the app open the same closed file (same name) and append the data
received. After that, the user should press the "STOP" button, so the (a)
situation will happen.