Set a hard-coded timeout for the measuring process
In the code of the nRF52 now is added a parameter to set a time-out
for the measuring task (is 60 seconds in this version). The idea is
that, for future versions, the this time-out is set by the android
app.

In this version was changed a bit the logic of the STAT characteristic
variables: commandFromPhone, isMeasuring, isDataOnFlash. The logic to
save the file though, works just like the past commit. However now, the
measuring process can be stopped by either the "STOP" button of the
android app or the hard-coded timeout of the nRF52.