Add hard-coded timeout from android app
Now, the data STAT characteristic has 4 more item: timeDuration_hrs,
timeDuration_mins, timeDuration_secs and tempReference. The timeDuration
is added to define the timeout of the measuring process which must be
defined by the user in the phone app at the beggining just before pressing
the "START" button. The tempReference is added but is not used in this
version, is meant to be used in the future when the PID controller is
implemented.

The flow is as follows: once nRF52 and android app are connected, the user
press the "START" button. Inside the code of the android app, as long as
the "STAT" variable isDataOnFlash is false, the sends to the nRF52 the
timeDuration and the tempReference by writing in the "STAT" characteristic.
So far, the timeDuration and the tempReference is hard-coded in the android
app, but for a future version this must be setteable by the user in the UI.

After the nRF52 receives the "STAT" characteristic from the android app
with the info of the timeDuration, then the measuring task is checking
if time is longer than the timeout value received to stop the measuring
process.

Note that the feature of saving the file is working succesfully with this
new timeout feature as in the two previous commits.