Fix bug closing file after disconnection
The past behavior of the android app was when the user disconnects,
then the file remains open and the data is not saved (it's saved after
closing, which is not a desirable behavior either).

Now, the user disconnects, then the file is closed, so the data is saved.
Note that now, pressing the "DISCONNECT" button producess the same
behavior as closing the android app.