Add pwm peripheral for pid controller
This commit just adds a simple functional pwm functionality. Note that
somo source files must be included in order to use pwm such ass 'ppi',
'timer' and 'app timer'. Note that the 'sdk_config.h' must set some
flags as usual. For more information, please compare ('git diff')
between this commit and the previous one.

This functional pwm code is based on the examples of the sdk located in
nRF5_SDK_16.0.0\examples\peripheral\pwm_library.

Note that the pin of the LED1 is used as the output of the pwm (overriding
the previous functionaly of turn on the LED for fluorescense), however
this pin can be easily changed to another pin for a future version.