Save on external flash temp control data
This version of the code enables to save data o the external flash and
then read for send over BLE. This just test the functionality, actually
doesn't send data via BLE, it just displays information on the serial
monitor instead.

The flow is the following: every second the nRF52 stores on the
external flash a control system data. After 400 counts (40 seconds), the
nRF52 read as fast as posible the data stored on flash a in printed via
serial monitor. Once all the data stored on the flash and in a internal
buffer is read, the nRF52 prints the last data stored until other 40
seconds. Then, the cycle is repeated.