Fix bug when saving on external flash
This is just tested partially, por it works fine for our purporses.
Bofere this fix, when the nRF52 read from the buffer (not the flash) the
flash was never erased, which is a problem when the head index reached
the end of the cycle. This bug was fixed for the control data and the
detection data.