Add CONT ble service
This version adds a ble service just like the SENS service with a
characteristic with the nitification property. This code can be tested
with the nRFconnect android app by enabling or disabling notifications,
you can see in the serial monitor how the nRF52 reacts and sends control
data to the phone. Note how the nRF52 is constantly saving data on the
external flash and the data is read and send after notifications are
enabled by the nRFconnect phone app.

Note that it was necessary to add a count for the number of ble services
in the "sdk_config.h" in the definition NRF_SDH_BLE_VS_UUID_COUNT. Also,
the RAM_START and RAM_SIZE in the "linker options" "Section Placement
Macros" were changed.