Add functions for status data ("sens" and "cont")
This functions are added to detect if there is data on flash saved in
the sectors of the "SENS" data or in the sectors of the "CONT" data. In
a near future, the logic of when to start the measuring process and how
much the phone have to wait for data stored on flash will have to change
a bit.

Also, some names and comments where changed for better readability.