Add "CONT" service in phone app
The "CONT" service/characteristic was added in the android code, the
code is similar to the "SENS" characteristic.

Also, the logic to detect when there is no data on flash is implemented
(now there is no data on flash as long as both there is no data due to
"Sens" and "Cont" data stored on flash).

It is possible to debug by looking the serial monitor of the nRF52 and
the logs printed out from android studio. In android studio it is
possible to see the notification data from the "CONT" characteristic.