Display "CONT" data on phone app
Now the data received from the "CONT" characteristic is displayed on the
phone app. Time, Reference, Adc and Pwm information from the controller
system is shown. You can see how the pid controller works for a RC
system.

Note that the pin of the pwm was changed to the (1.15) arduino pin 13.