Read state from battery babysitter
This is a simple demostration which implements communication between
the battery babysitter and the nRF52 via TWI (or I2C). Basically, this
code version adds functions for twi communication to get data such as
soc, soh, voltage, current, power, capacity, etc.

Code related with setting the default capacity of the battery was left
commented out for a future implementation since in this try the code for
this purpose didn't work :(.

Also, it was detected a possible bug for twi communication with the ads1015.
The solution of this bug was commented with a "TODO" tag for a future
analysis.