Add timer for battery system
This version simply adds a timer for the battery system. This is
implemented in the nRF52 and can be debbuged by seeing the
serial communication where is displayed the data read from the
battery babysitter.

Note that the following lines where necessary after all:
//TODO: check if the two following lines are necessary.
//while(!m_xfer_done);
//m_xfer_done = false;
otherwise the nRF52 gets stucked due to twi-i2c communication.