Add functions for status data ("batt")
This functions are added to detect if there is data on flash saved in
the sectors of the "BATT" data. In a near future, the logic of when to
start the measuring process and how much the phone have to wait for data
stored on flash will have to change a bit.

It is possible to test this code with the nRFconnect android app. In the
nRFconnect it's necessary to enable notifications of the "STAT" characteristic
firts and then enable or disable notifications of the "BATT" characteristic.
After that, it's necessary to write in in the "STAT" characteristic the byte
0x01. You can see in the serial monitor how the nRF52 reacts and sends battery
data to the phone. Note how the nRF52 is constantly saving data on the
external flash and the data is read and send after notifications are enabled
by the nRFconnect phone app.