Add "BATT" service in phone app
The "BATT" service/characteristic was added in the android code, the
code is similar to the "CONT" characteristic.

Also, the logic to detect when there is no data on flash is implemented
(now there is no data on flash as long as there is no data due to
"Sens", "Cont" and "Batt" data stored on flash).

Note that in the nRF52 when the process of measuring is running, the
battery data is stored on flash and send via BLE. When the nRF52 is not
measuring, it sends battery data as long as the phone is connected with
BATT notifications are enabled.

It is possible to debug by looking the serial monitor of the nRF52 and
the logs printed out from android studio. In android studio it is
possible to see the notification data from the "BATT" characteristic.