Add activity structure for plotting from files
This version adds two activities: one for displaying the files stored in
the DNAamplifier folder and the other one for display the data stored
one the selected file. Up to now, the file is not read, however the
structure is setup for a future implementation for plotting from the
saved files.