Plot SENS, CONT and BATT data in the PlotActivity
Now it's plotted the data saved in the files "_measure.txt",
"_control.txt" and "_battery.txt" in the PlotActivity. Note that there
was a little bug, when the nordic sent the last battery data, its time
was 0, this was changed in the nRF52 code by simply changing the time
value at the end when the "seconds" timer is stopped.