Changes for feather nRF52 hardware
Some pin assignation was changes. Be Aware! some changes of the pins
were made in the pca10056 file wich is not included in this git project.

Also, 4 unit16_t variables were added for the controller system:
refAdcHotlid, adcValueHotlid, pwmValueHotlid and adcTamb. Thes variables
were obtained by adding some ADC inputs and PWM outputs. Note that these
variable are stored in the flash and sent via BLE.

The android app was changed a bit to display thes new 4 variables.

Finally, now are 2 PID temperature controllers. One for heating the
samples and another one for controling the temperature if the hotlid.